define the producer and separate it from the chain
