+++ pkg/core/chain/catcher.go
@@ -1,0 +1,1 @@
+package chain
